chore(jackson)!: json modules restructuring
