+++ connector-sdk/jackson-datatype-feel/pom.xml
@@ -17,23 +17,6 @@
       <artifactId>connector-feel-wrapper</artifactId>
     </dependency>
 
-    <!-- additional Jackson modules for the pre-configured ObjectMapper supplier -->
-    <dependency>
-      <groupId>com.fasterxml.jackson.datatype</groupId>
-      <artifactId>jackson-datatype-jsr310</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.datatype</groupId>
-      <artifactId>jackson-datatype-jdk8</artifactId>
-    </dependency>
-
-    <!-- brings @FEEL annotation -->
-    <dependency>
-      <groupId>io.camunda.connector</groupId>
-      <artifactId>connector-core</artifactId>
-      <optional>true</optional>
-    </dependency>
-
     <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter</artifactId>
