+++ connector-sdk/jackson-datatype-feel/src/main/java/io/camunda/connector/feel/jackson/JacksonModuleFeelFunction.java
@@ -46,6 +46,7 @@ public class JacksonModuleFeelFunction extends SimpleModule {
     addDeserializer(
         Supplier.class,
         new FeelSupplierDeserializer<>(TypeFactory.unknownType(), feelEngineWrapper));
+    context.insertAnnotationIntrospector(new FeelAnnotationIntrospector());
     super.setupModule(context);
   }
 }
