+++ connectors/http/http-base/src/main/java/io/camunda/connector/http/base/auth/OAuthAuthentication.java
@@ -17,7 +17,7 @@
 package io.camunda.connector.http.base.auth;
 
 import com.google.api.client.http.HttpHeaders;
-import io.camunda.connector.api.annotation.FEEL;
+import io.camunda.connector.feel.annotation.FEEL;
 import io.camunda.connector.generator.annotation.TemplateProperty;
 import io.camunda.connector.generator.annotation.TemplateProperty.DropdownPropertyChoice;
 import io.camunda.connector.generator.annotation.TemplateProperty.PropertyType;
