+++ connectors/http/http-base/src/test/java/io/camunda/connector/http/base/services/HttpInteractionServiceTest.java
@@ -26,7 +26,7 @@ import static org.mockito.Mockito.when;
 
 import com.google.api.client.http.HttpHeaders;
 import com.google.api.client.http.HttpResponse;
-import io.camunda.connector.feel.ConnectorsObjectMapperSupplier;
+import io.camunda.connector.api.json.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.http.base.model.HttpCommonResult;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
