+++ connectors/http/polling/src/main/java/io/camunda/connector/http/polling/HttpPollingConnector.java
@@ -9,7 +9,7 @@ package io.camunda.connector.http.polling;
 import io.camunda.connector.api.annotation.InboundConnector;
 import io.camunda.connector.api.inbound.InboundConnectorExecutable;
 import io.camunda.connector.api.inbound.InboundIntermediateConnectorContext;
-import io.camunda.connector.feel.ConnectorsObjectMapperSupplier;
+import io.camunda.connector.api.json.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.http.base.components.HttpTransportComponentSupplier;
 import io.camunda.connector.http.base.services.HttpService;
 import io.camunda.connector.http.polling.service.SharedExecutorService;
