+++ connectors/http/polling/src/test/java/io/camunda/connector/http/polling/model/PollingIntervalConfigurationTest.java
@@ -9,8 +9,8 @@ package io.camunda.connector.http.polling.model;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
+import io.camunda.connector.api.json.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.api.secret.SecretProvider;
-import io.camunda.connector.feel.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.runtime.core.inbound.InboundConnectorContextImpl;
 import io.camunda.connector.runtime.core.inbound.InboundConnectorDefinitionImpl;
 import java.util.HashMap;
