+++ connectors/http/rest/src/main/java/io/camunda/connector/http/rest/HttpJsonFunction.java
@@ -19,9 +19,9 @@ package io.camunda.connector.http.rest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.api.client.http.HttpRequestFactory;
 import io.camunda.connector.api.annotation.OutboundConnector;
+import io.camunda.connector.api.json.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.api.outbound.OutboundConnectorContext;
 import io.camunda.connector.api.outbound.OutboundConnectorFunction;
-import io.camunda.connector.feel.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.generator.annotation.ElementTemplate;
 import io.camunda.connector.generator.annotation.ElementTemplate.PropertyGroup;
 import io.camunda.connector.http.base.components.HttpTransportComponentSupplier;
