+++ connectors/rabbitmq/src/main/java/io/camunda/connector/rabbitmq/supplier/ObjectMapperSupplier.java
@@ -7,7 +7,7 @@
 package io.camunda.connector.rabbitmq.supplier;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.camunda.connector.feel.ConnectorsObjectMapperSupplier;
+import io.camunda.connector.api.json.ConnectorsObjectMapperSupplier;
 
 public final class ObjectMapperSupplier {
