+++ connectors/rabbitmq/src/test/java/io/camunda/connector/rabbitmq/BaseTest.java
@@ -11,7 +11,7 @@ import static java.nio.file.Files.readString;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import io.camunda.connector.feel.ConnectorsObjectMapperSupplier;
+import io.camunda.connector.api.json.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.rabbitmq.supplier.ObjectMapperSupplier;
 import java.io.File;
 import java.io.IOException;
